feat: baseEquals 함수 추가 및 shallowEquals, deepEquals 함수에서 사용
